requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
pillow
python-multipart
